SCC-41 query assembler decomposition
