+++ src/query_assembler_ddl.cpp
@@ -1,74 +1,5 @@
 #include "SCC/query_assembler.h"
 
-void QueryAssembler::Translate(std::shared_ptr<Node> AST) {
-  LOG(INFO, "Starting translation...");
-
-  ast_ = std::move(AST);
-
-  if (ast_ == nullptr) {
-    LOG(TRACE, "empty AST, nothing to translate");
-    LOG(INFO, "Translation is ended");
-    return;
-  }
-
-  if (ast_->get_st_type() == StatementType::Program) {
-    if (ast_->get_children_amount() > 0) {
-      this->TranslateProgram(ast_);
-    } else {
-      LOG(TRACE, "root of AST does not have children");
-      LOG(INFO, "Translation is ended");
-      return;
-    }
-  } else {
-    LOG(ERROR,
-        "invalid AST: root should be with \'Program\' statement type");
-    exit(EXIT_FAILURE);
-  }
-
-  LOG(INFO, "Translation is ended");
-}
-
-void QueryAssembler::TranslateProgram(std::shared_ptr<Node> node) {
-  auto query = node->get_child(0);
-  if (query->get_st_type() == StatementType::query) {
-    this->TranslateQuery(query);
-  } else {
-    LOG(ERROR, "first child of program node is not a query");
-    exit(EXIT_FAILURE);
-  }
-
-  if (node->get_children_amount() > 1) {
-    auto other_queries = node->get_child(1);
-    if (other_queries->get_st_type() == StatementType::delimiter_semicolon) {
-      if (other_queries->get_children_amount() > 0) {
-        this->TranslateProgram(other_queries);
-      }
-    } else {
-      LOG(ERROR, "invalid delimiter between queries");
-      exit(EXIT_FAILURE);
-    }
-  }
-}
-void QueryAssembler::TranslateQuery(std::shared_ptr<Node> node) {
-  if (node->get_children_amount() == 0) {
-    LOG(TRACE, "empty query");
-    return;
-  }
-
-  auto data_language = node->get_child(0);
-  switch (data_language->get_st_type()) {
-    case StatementType::ddlStatement:
-      this->TranslateDDLStatement(data_language);
-      break;
-    case StatementType::dmlStatement:
-      this->TranslateDMLStatement(data_language);
-      break;
-    default:
-      LOG(ERROR, "unknown query data language");
-      exit(EXIT_FAILURE);
-  }
-}
-
 void QueryAssembler::TranslateDDLStatement(std::shared_ptr<Node> node) {
   if (node->get_children_amount() == 0) {
     LOG(TRACE, "empty DDL query");
@@ -94,27 +25,6 @@ void QueryAssembler::TranslateDDLStatement(std::shared_ptr<Node> node) {
       break;
     default:
       LOG(ERROR, "unknown DDL statement");
-      exit(EXIT_FAILURE);
-  }
-}
-void QueryAssembler::TranslateDMLStatement(std::shared_ptr<Node> node) {
-  if (node->get_children_amount() == 0) {
-    LOG(TRACE, "empty DML query");
-  }
-
-  auto statement = node->get_child(0);
-  switch (statement->get_st_type()) {
-    case StatementType::insertStatement:
-      this->TranslateInsert(statement);
-      break;
-    case StatementType::deleteStatement:
-      this->TranslateDelete(statement);
-      break;
-    case StatementType::updateStatement:
-      this->TranslateUpdate(statement);
-      break;
-    default:
-      LOG(ERROR, "unknown DML statement");
       exit(EXIT_FAILURE);
   }
 }
@@ -517,235 +427,4 @@ void QueryAssembler::TranslateDropObject(
     LOG(ERROR, "unknown type of the drop object");
     exit(EXIT_FAILURE);
   }
-}
-
-void QueryAssembler::TranslateInsert(std::shared_ptr<Node> node) {}
-void QueryAssembler::TranslateDelete(std::shared_ptr<Node> node) {}
-void QueryAssembler::TranslateUpdate(std::shared_ptr<Node> node) {}
-
-// Basic statements
-
-void QueryAssembler::TranslatePrimaryKey(
-    std::shared_ptr<Node> key,
-    std::string& constraint_name,
-    std::string& table_name) {
-  if (key->get_st_type() != StatementType::primaryKey) {
-    LOG(ERROR, "incorrect type for primaryKey node");
-    exit(EXIT_FAILURE);
-  }
-  if (key->get_children_amount() == 0) {
-    LOG(ERROR, "PRIMARY KEY definition is missed");
-    exit(EXIT_FAILURE);
-  }
-
-  std::vector<std::string> properties;
-  properties.push_back(this->TranslateIdentifier(key->get_child(0)));
-  if (key->get_children_amount() > 1) {
-    std::vector<std::string> other_properties =
-        this->GetListOf(key->get_child(1), StatementType::identifier);
-    properties.insert(properties.end(),
-                      other_properties.begin(),
-                      other_properties.end());
-  }
-
-  this->CreateUniqueNodePropertyConstraint(
-      (constraint_name + "_" + std::to_string(constraint_counter++)),
-      table_name,
-      properties
-  );
-  for (auto& i: properties) {
-    this->CreateNodePropertyExistenceConstraint(
-        (constraint_name + "_" + std::to_string(constraint_counter++)),
-        table_name,
-        i
-    );
-  }
-}
-void QueryAssembler::TranslateForeignKey(
-    std::shared_ptr<Node> key,
-    std::string& table_name) {
-  if (key->get_st_type() != StatementType::foreignKey) {
-    LOG(ERROR, "incorrect type for foreignKey node");
-    exit(EXIT_FAILURE);
-  }
-  if (key->get_children_amount() == 0) {
-    LOG(ERROR, "PRIMARY KEY definition is missed");
-    exit(EXIT_FAILURE);
-  }
-  int reference_child_num = 1;
-
-  std::vector<std::string> properties;
-  properties.push_back(this->TranslateIdentifier(key->get_child(0)));
-  if (key->get_children_amount() > 2) {
-    if (key->get_child(1)->get_st_type() != StatementType::delimiter_comma) {
-      LOG(ERROR, "invalid delimiter between properties in foreign key");
-      exit(EXIT_FAILURE);
-    }
-    reference_child_num++;
-    std::vector<std::string> other_properties =
-        this->GetListOf(key->get_child(1), StatementType::identifier);
-    properties.insert(properties.end(),
-                      other_properties.begin(),
-                      other_properties.end());
-  }
-
-  // Get reference
-  auto reference = key->get_child(reference_child_num);
-  if (reference->get_st_type() != StatementType::reference) {
-    LOG(ERROR, "invalid foreign key: incorrect referene statement type");
-    exit(EXIT_FAILURE);
-  }
-  if (reference->get_children_amount() == 0) {
-    LOG(ERROR, "empty reference");
-    exit(EXIT_FAILURE);
-  }
-  auto table_name_node = reference->get_child(0);
-  std::string ref_table_name = this->TranslateName(table_name_node);
-
-  // Get ref columns if present
-  std::vector<std::string> ref_columns;
-  if (reference->get_children_amount() > 1) {
-    ref_columns.push_back(this->TranslateIdentifier(reference->get_child(1)));
-    if (reference->get_children_amount() > 2) {
-      std::vector<std::string> other_props =
-          this->GetListOf(reference->get_child(2),
-                          StatementType::identifier);
-      ref_columns.insert(ref_columns.end(),
-                         other_props.begin(),
-                         other_props.end());
-    }
-  }
-
-  this->RemoveProperties(table_name, properties);
-  this->RemoveProperties(table_name, ref_columns);
-  this->CreateRelationship(table_name, ref_table_name);
-}
-void QueryAssembler::CreateUniqueNodePropertyConstraint(
-    const std::string& constraint_name,
-    const std::string& LabelName,
-    const std::vector<std::string>& properties) {
-  out_ << "CREATE CONSTRAINT [" << constraint_name << "]" << std::endl;
-  out_ << "FOR (n:" << LabelName << ")" << std::endl;
-  out_ << "REQUIRE (";
-  for (size_t i = 0; i < properties.size(); i++) {
-    out_ << "n." << properties[i];
-    if (i + 1 != properties.size()) {
-      out_ << ", ";
-    }
-  }
-  out_ << ") IS UNIQUE;\n" << std::endl;
-}
-void QueryAssembler::CreateNodePropertyExistenceConstraint(
-    const std::string& constraint_name,
-    const std::string& LabelName,
-    const std::string& property) {
-  out_ << "CREATE CONSTRAINT [" << constraint_name << "]" << std::endl;
-  out_ << "FOR (n:" << LabelName << ")" << std::endl;
-  out_ << "REQUIRE (n." << property << ") IS NOT NULL;\n" << std::endl;
-}
-void QueryAssembler::CreateRelationship(
-    const std::string& label_name,
-    const std::string& ref_label_name) {
-  out_ << "MATCH (a:" << label_name << "), (b:" << ref_label_name << ")\n";
-  out_ << "CREATE (a)-[r:fk_"
-       << label_name << "_to_" << ref_label_name << "_" << relationship_counter
-       << "]->(b)\n" << std::endl;
-  relationship_counter++;
-}
-void QueryAssembler::RemoveProperties(
-    const std::string& label_name,
-    const std::vector<std::string>& properties) {
-  if (properties.empty()) {
-    return;
-  }
-  out_ << "MATCH (n:" << label_name << ")\n";
-  out_ << "SET ";
-  for (size_t i = 0; i < properties.size(); i++) {
-    out_ << "n." << properties[i] << " = null";
-    if (i + 1 != properties.size()) {
-      out_ << ", ";
-    }
-  }
-  out_ << ";\n" << std::endl;
-}
-
-std::vector<std::string> QueryAssembler::GetListOf(
-    std::shared_ptr<Node> node,
-    StatementType type) {
-  if (node->get_st_type() != StatementType::delimiter_comma) {
-    LOG(ERROR, "invalid ListOf: delimiter is not a comma");
-    exit(EXIT_FAILURE);
-  }
-  if (node->get_children_amount() == 0) {
-    LOG(ERROR, "invalid ListOf: comma without children");
-    exit(EXIT_FAILURE);
-  }
-
-  std::vector<std::string> arguments;
-  switch (type) {
-    case StatementType::name:
-      arguments.push_back(this->TranslateName(node->get_child(0)));
-      break;
-    case StatementType::identifier:
-      arguments.push_back(this->TranslateIdentifier(node->get_child(0)));
-      break;
-    default:
-      LOG(ERROR, "invalid ListOf: unknown argument type");
-      exit(EXIT_FAILURE);
-  }
-
-  if (node->get_children_amount() > 1) {
-    std::vector<std::string> other_arguments =
-        this->GetListOf(node->get_child(1), type);
-    arguments.insert(arguments.end(),
-                     other_arguments.begin(),
-                     other_arguments.end());
-  }
-
-  return arguments;
-}
-
-std::string QueryAssembler::TranslateName(std::shared_ptr<Node> node) {
-  if (node->get_children_amount() == 0) {
-    LOG(ERROR, "empty name");
-    exit(EXIT_FAILURE);
-  }
-
-  std::ostringstream name;
-
-  name << this->TranslateIdentifier(node->get_child(0));
-  if (node->get_children_amount() > 1) {
-    if (node->get_child(1)->get_st_type() == StatementType::delimiter_dot) {
-      name << this->TranslateIdentifiers(node->get_child(1));
-    } else {
-      LOG(ERROR, "invalid name: incorrect delimiter");
-      exit(EXIT_FAILURE);
-    }
-  }
-
-  return name.str();
-}
-std::string QueryAssembler::TranslateIdentifiers(
-    std::shared_ptr<Node> node) {
-  if (node->get_children_amount() == 0) {
-    LOG(ERROR, "invalid list of identifiers");
-    exit(EXIT_FAILURE);
-  }
-
-  std::ostringstream identifiers;
-  identifiers << "." << this->TranslateIdentifier(node->get_child(0));
-
-  if (node->get_children_amount() > 1) {
-    identifiers << this->TranslateIdentifiers(node->get_child(1));
-  }
-
-  return identifiers.str();
-}
-std::string QueryAssembler::TranslateIdentifier(std::shared_ptr<Node> node) {
-  if (node->get_children_amount() == 0) {
-    LOG(ERROR, "empty identifier");
-    exit(EXIT_FAILURE);
-  }
-
-  return std::dynamic_pointer_cast<StringNode>(node->get_child(0))->get_data();
 }
